requirements.txt
streamlit>=1.32.0
supabase>=2.4.0
python-dotenv>=1.0.0
pandas>=2.2.0
openpyxl>=3.1.2
reportlab>=4.1.0
